+++ apps/betterangels-backend/clients/tests/test_models.py
@@ -1,0 +1,30 @@
+from clients.enums import HmisAgencyEnum, PronounEnum
+from clients.models import ClientProfile, HmisProfile
+from django.db import IntegrityError
+from django.test import TestCase
+from model_bakery import baker
+
+
+class ClientProfileTestCase(TestCase):
+    def test_display_pronouns(self) -> None:
+        client_profile = baker.make(ClientProfile)
+        self.assertIsNone(client_profile.display_pronouns)
+
+        client_profile.pronouns = PronounEnum.HE_HIM_HIS
+        client_profile.save()
+        self.assertEqual(client_profile.display_pronouns, "He/Him/His")
+
+        client_profile.pronouns = PronounEnum.OTHER
+        client_profile.pronouns_other = "she/her/their"
+        client_profile.save()
+        self.assertEqual(client_profile.display_pronouns, "she/her/their")
+
+
+class HmisProfileModelTestCase(TestCase):
+    def test_hmis_profile_unique_constraint(self) -> None:
+        baker.make(HmisProfile, hmis_id="hmisID1", agency=HmisAgencyEnum.LAHSA)
+        baker.make(HmisProfile, hmis_id="hmisID2", agency=HmisAgencyEnum.LAHSA)
+        baker.make(HmisProfile, hmis_id="hmisID1", agency=HmisAgencyEnum.PASADENA)
+
+        with self.assertRaises(IntegrityError):
+            baker.make(HmisProfile, hmis_id="hmisID1", agency=HmisAgencyEnum.LAHSA)
